making ConfigJSON local
While trying to do the multi-instance verson of cspot, I hit the wall that COnfigJSON is a global item. This addition to the PR makes it local. It is still shared by many entities, maybe too many, but I've tried to make it available on a "need to know basis" and when entities can dynamically see config changes. So for example, a track does not need to see changes of format, it can be given at creation of object